prestasi_mahasiswa/views.py
from django.shortcuts import render
from rest_framework.decorators import api_view
from rest_framework.response import Response

@api_view(['POST'])
def insertForm(request):
    return Response(request.data)