requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
decorator==4.0.11
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-debugtools==2.0
djangorestframework==3.13.1
filelock==3.7.0
filetype==1.0.5
idna==3.3
imageio==2.1.2
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instapy-cli==0.0.12
moviepy==0.2.3.2
numpy==1.22.0
Pillow==9.0.0
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.11.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.14.1
